MAINT: Move the `fromnumeric` annotations to their own stub file
